requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
Automat==22.10.0
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.29.7
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
cftime==1.6.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
configparser==6.0.0
constantly==23.10.4
contourpy==1.0.7
cpp==1.0.0.6
cryptography==42.0.3
cssselect==1.2.0
cycler==0.11.0
DateTime==5.5
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Django==5.0.2
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.13.1
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
gTTS==2.3.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
isodate==0.6.1
isoduration==20.11.0
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterthemes==0.20.0
kiwisolver==1.4.4
lesscpy==0.15.1
lxml==5.1.0
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
memory-profiler==0.61.0
mistune==2.0.5
mpg123==0.4
multidict==6.0.4
mysql-connector==2.2.9
mysql-connector-python==8.0.32
mysqlclient==2.2.4
Naked==0.1.32
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.0
nbformat==5.8.0
nest-asyncio==1.5.6
netCDF4==1.7.1
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.24.1
oauthlib==3.2.2
openai==0.27.2
opencv-python==4.10.0.82
openpyxl==3.1.2
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.0
parsel==1.8.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.2.0
playsound==1.3.0
plotly==5.14.1
ply==3.11
prometheus-client==0.16.0
prompt-toolkit==3.0.38
Protego==0.3.0
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
PyDispatcher==2.0.7
pygame==2.1.2
Pygments==2.14.0
pyodbc==4.0.35
pyOpenSSL==24.0.0
pyparsing==3.0.9
pypng==0.20220715.0
pyrsistent==0.19.3
pyshorteners==1.0.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2022.7.1
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qrcode==7.4.2
queuelib==1.6.2
requests==2.28.2
requests-file==2.0.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-learn==1.2.2
scipy==1.10.1
Scrapy==2.11.1
seaborn==0.12.2
Send2Trash==1.8.0
service-identity==24.1.0
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
sqlparse==0.4.4
stack-data==0.6.2
tenacity==8.2.2
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tk==0.1.0
tldextract==5.1.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
tweepy==4.14.0
Twisted==23.10.0
twisted-iocpsupport==1.0.4
typing_extensions==4.9.0
tzdata==2024.1
uri-template==1.2.0
urllib3==1.26.15
w3lib==2.1.2
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
xarray==2024.6.0
yarl==1.8.2
zope.interface==6.2
